+++ app/routes/web.php
@@ -244,6 +244,8 @@ $router->group(['middleware' => 'auth'], static function (Router $router) {
             $router->get('misc/restart')->uses('Dominion\MiscController@getRestartDominion')->name('misc.restart');
             $router->post('misc/restart')->uses('Dominion\MiscController@postRestartDominion');
             $router->post('misc/rename')->uses('Dominion\MiscController@postRenameDominion')->name('misc.rename');
+            $router->get('misc/settings')->uses('Dominion\MiscController@getDominionSettings')->name('misc.settings');
+            $router->post('misc/settings')->uses('Dominion\MiscController@postDominionSettings');
             $router->get('misc/tick')->uses('Dominion\MiscController@getTickDominion')->name('misc.tick');
             $router->get('misc/undo-tick')->uses('Dominion\MiscController@getUndoTickDominion')->name('misc.undo-tick');